rf: move variance into its own file
